requirements.txt
asgiref==3.7.2
boto3==1.26.152
botocore==1.29.152
Django==4.2.1
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-editorjs-fields==0.2.7
django-environ==0.10.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-storages==1.13.2
djangorestframework==3.14.0
fake-factory==9999.9.9
Faker==18.10.1
gunicorn==20.1.0
jmespath==1.0.1
model-bakery==1.12.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==1.26.16
whitenoise==6.4.0